worker/requirements.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
numpy>=1.26.0
scikit-learn>=1.4.0
hdbscan>=0.8.33
python-dotenv>=1.0.0
anthropic>=0.25.0
feedparser>=6.0.11
